pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Daniil Petrov <<xjem666@yandex.ru>>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"


[tool.poetry.scripts]
gendiff = 'gendiff.gendiff:main'


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"
PyYAML = "^6.0"
pytest = "^7.4.3"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
